fixup: Complete refactor move of FeatureExtractor
